Update README.md
When there's no any configuration in `new WorkerPlugin()`, there would be such a warning message:
```
Warning (worker-plugin): output.globalObject is set to "window". It must be set to "self" to support HMR in Workers.
```
So It seems the default value is `window`, and it's incorrect in the `README.md` reads that the default value is `self`.